refactor(replay): split files under backend size limits
